+++ crates/core/action-pipeline/src/actions/sync_workspace.rs
@@ -1,7 +1,7 @@
 use super::should_skip_action;
 use moon_action::{Action, ActionStatus};
 use moon_action_context::ActionContext;
-use moon_actions::{install_proto, sync_codeowners, sync_vcs_hooks};
+use moon_actions::{sync_codeowners, sync_vcs_hooks};
 use moon_logger::debug;
 use moon_project_graph::ProjectGraph;
 use moon_utils::is_test_env;
@@ -27,10 +27,6 @@ pub async fn sync_workspace(
     let workspace = workspace.read().await;
     let project_graph = project_graph.read().await;
 
-    // Install proto
-    install_proto(&workspace).await?;
-
-    // Sync workspace
     debug!(target: LOG_TARGET, "Syncing workspace");
 
     if should_skip_action("MOON_SKIP_SYNC_WORKSPACE") {